pyproject.toml
[project]
name = "finance-track"
version = "0.1.0"
description = ""
authors = [
    {name = "Antonio Júnior",email = "juniorssz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
